refactor code and organize files
